pyproject.toml
[tool.poetry]
name = "pythoncitygmlrequest"
version = "0.1.0"
description = ""
authors = ["Sensyn Robotics"]

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
cjio = "0.7.3"
pyproj = "3.2.1"
PyYAML = "6.0"
Shapely = "1.8.2"
trimesh = "3.12.6"
open3d = "0.15.2"
pymeshlab = "^2022.2.post2"
pyjpmesh = "^1.1.0"
contexttimer = "^0.3.3"
pyglet = "^1.5.26"
Rtree = "^1.0.0"
mapbox-earcut = "^1.0.1"
networkx = "^3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
